+++ vortex-duckdb/src/multi_file.rs
@@ -21,8 +21,8 @@ use vortex_utils::aliases::hash_map::HashMap;
 use crate::RUNTIME;
 use crate::SESSION;
 use crate::datasource::DataSourceTableFunction;
-use crate::datasource::HivePartitionColumn;
-use crate::datasource::extract_hive_partition_columns;
+use crate::hive::HivePartitionColumn;
+use crate::hive::extract_hive_partition_columns;
 use crate::duckdb::BindInputRef;
 use crate::duckdb::ClientContextRef;
 use crate::duckdb::ExtractedValue;
